requirements.txt
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
colour==0.1.5
credentials==1.1
debugpy==1.5.1
decorator==5.1.0
entrypoints==0.3
idna==3.3
imageio==2.10.3
ipykernel==6.5.0
ipython==7.29.0
jedi==0.18.0
joblib==1.1.0
jupyter-client==7.0.6
jupyter-core==4.9.1
langdetect==1.0.9
lxml==4.6.4
matplotlib-inline==0.1.3
nest-asyncio==1.5.1
nltk==3.6.5
numpy==1.21.4
oauthlib==3.1.1
osmviz==3.4.1
pandas==1.3.4
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.22
Pygments==2.10.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
pywin32==302
pyzmq==22.3.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
scikit-learn==1.0.1
scipy==1.7.2
six==1.16.0
textblob==0.17.1
threadpoolctl==3.0.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
tweepy==3.9.0
urllib3==1.26.7
wcwidth==0.2.5
